+++ src/accessories/childlock/childlock.accessory.js
@@ -29,10 +29,7 @@ class Accessory {
       );
     }
 
-    service
-      .getCharacteristic(this.api.hap.Characteristic.On)
-      .onGet(() => this.handler.get(this.accessory))
-      .onSet((state) => this.handler.set(state, this.accessory));
+    service.getCharacteristic(this.api.hap.Characteristic.On).onSet((state) => this.handler.set(state, this.accessory));
   }
 }